Extract list verification to editorActions
