+++ apps/tangent-electron/src/app/views/editors/NoteEditor/editorModule.ts
@@ -37,6 +37,7 @@ import type MarkdownEditor from './MarkdownEditor'
 import { appendContextTemplate, type ContextMenuConstructorOptions } from 'app/model/menus'
 import { eventHasSelectionRequest } from 'app/events'
 import { createCommandHandler } from 'app/model/commands/Command'
+import { verifyListContext } from './editorActions'
 
 function clampRange(range: EditorRange, clampingRange: EditorRange): EditorRange {
 	range = normalizeRange(range)
@@ -50,15 +51,6 @@ function clampRange(range: EditorRange, clampingRange: EditorRange): EditorRange
 interface VerificationInstruction<T> {
 	func: (options: T, change: TextChange) => TextChange,
 	options: T
-}
-
-interface VerifyListOptions {
-	id: string,
-	targetIndent: string,
-	// Whether to apply the line's list format or incorporate into siblings'
-	basis: 'self' | 'rebasis'
-	// Whether to enforce that unordered glyphs are the same
-	normalizeUnorderedGlyphs: boolean
 }
 
 // Force the inclusion of elements so it is included in the module
@@ -268,33 +260,39 @@ export default function editorModule(editor: Editor, options: {
 				func: verifyListContext,
 				options: {
 					id,
+					editor: markdownEditor,
 					targetIndent: newIndent,
 					basis: (isGlyphCreator && newList) ? 'self' : 'rebasis',
-					normalizeUnorderedGlyphs: true
+					normalizeUnorderedGlyphs: true,
+					autoSetChildGlyphs: workspace?.settings.autoSetChildListGlyphs.value
 				}
 			})
 			pushVerification({
 				func: verifyListContext,
 				options: {
 					id,
+					editor: markdownEditor,
 					targetIndent: oldIndent,
 					basis: 'rebasis',
-					normalizeUnorderedGlyphs: true
+					normalizeUnorderedGlyphs: true,
+					autoSetChildGlyphs: workspace?.settings.autoSetChildListGlyphs.value
 				}
 			})
 		} 
 		else if (oldList && newList) {
 			if (!ListDefinition.areEqualExceptForActiveTodoState(oldList, newList)) {
 				// Only do anything if there is a difference
-				if (oldList.indent === newList.indent) {
+				if (oldIndent === newIndent) {
 					// Any change to form should propegate down
 					pushVerification({
 						func: verifyListContext,
 						options: {
 							id,
+							editor: markdownEditor,
 							targetIndent: oldIndent,
 							basis: 'self',
-							normalizeUnorderedGlyphs: true
+							normalizeUnorderedGlyphs: true,
+							autoSetChildGlyphs: workspace?.settings.autoSetChildListGlyphs.value
 						}
 					})
 				}
@@ -306,9 +304,11 @@ export default function editorModule(editor: Editor, options: {
 				func: verifyListContext,
 				options: {
 					id,
+					editor: markdownEditor,
 					targetIndent: oldIndent,
 					basis: 'rebasis',
-					normalizeUnorderedGlyphs: isListGlyphCreationChange(delta) !== '\n'
+					normalizeUnorderedGlyphs: isListGlyphCreationChange(delta) !== '\n',
+					autoSetChildGlyphs: workspace?.settings.autoSetChildListGlyphs.value
 				}
 			})
 		}
@@ -322,211 +322,15 @@ export default function editorModule(editor: Editor, options: {
 					func: verifyListContext,
 					options: {
 						id,
+						editor: markdownEditor,
 						targetIndent: newIndent,
 						basis: isGlyphCreator ? 'self' : 'rebasis',
-						normalizeUnorderedGlyphs: true
+						normalizeUnorderedGlyphs: true,
+						autoSetChildGlyphs: workspace?.settings.autoSetChildListGlyphs.value
 					}
 				})
 			}
 		}
-	}
-
-	function verifyListContext(
-		options: VerifyListOptions,
-		change?: TextChange // An existing change to work with 
-	): TextChange {
-		const { doc } = editor
-		let selection = doc.selection?.slice()
-		let targetLine = doc.getLineBy(options.id)
-
-		if (!targetLine) return change
-
-		let lineText = deltaToText(targetLine.content)
-		let intendedIndent = options.targetIndent
-		
-		let targetRange = doc.getLineRange(targetLine)
-		let targetLineIndex = doc.lines.indexOf(targetLine)
-		
-		let targetListData = targetLine.attributes.list as ListDefinition
-		let targetIndent = targetLine.attributes.indent?.indent || ''
-
-		let targetForm: ListForm = undefined
-		let targetGlyph: string = undefined
-		let basisNumber: number = undefined
-
-		if (options.basis === 'self') {
-			if (!targetListData) {
-				console.error('Was told to verify a basis of "self", but the target line had no list data')
-				return change
-			}
-
-			targetForm = targetListData.form
-			targetGlyph = targetListData.glyph
-			basisNumber = targetListData.index
-		}
-		else if (!targetListData || targetIndent === intendedIndent || targetIndent.length > intendedIndent.length) {
-			// Find the basis for this indent level
-			for (let lineIndex = targetLineIndex - 1; lineIndex >= 0; lineIndex--) {
-				let prevLine = doc.lines[lineIndex]
-				let prevText = deltaToText(prevLine.content)
-				let indent = prevText.match(indentMatcher)[0]
-
-				if (indent === intendedIndent) {
-					// This is what we're looking for
-					const listData = prevLine.attributes.list as ListDefinition
-					if (listData) {
-						targetForm = listData.form
-						targetGlyph = listData.glyph
-						basisNumber = listData.index
-
-						if (basisNumber) basisNumber++
-					}
-					break
-				}
-				else if (indent.length < intendedIndent.length) {
-					if (lineIndex === targetLineIndex - 1 && (workspace?.settings.autoSetChildListGlyphs.value ?? true)) {
-						const listData = prevLine.attributes.list as ListDefinition
-						if (listData) {
-							const { form, glyph, basis } = getAutoChild(listData)
-							targetForm = form
-							targetGlyph = glyph
-							basisNumber = basis
-						}
-					}
-					break
-				}
-			}
-		}
-
-		if (targetForm === undefined || targetIndent.length < intendedIndent.length || (targetIndent === intendedIndent && !targetListData)) {
-			// Still don't have a target form. Look forward to find one
-			for (let lineIndex = targetLineIndex + 1; lineIndex < doc.lines.length; lineIndex++) {
-				let nextLine = doc.lines[lineIndex]
-				let nextText = deltaToText(nextLine.content)
-				let indent = nextText.match(indentMatcher)[0]
-	
-				if (indent === intendedIndent) {
-					const listData = nextLine.attributes.list as ListDefinition
-					if (!listData) break // End of the road
-	
-					// We might have removed the top out
-					targetForm = listData.form
-					targetGlyph = listData.glyph
-					basisNumber = listData.index
-
-					if (basisNumber) {
-						// Reset the basis number
-						basisNumber = 1
-					}
-					break
-				}
-				else if (indent.length < intendedIndent.length) {
-					break
-				}
-			}
-		}
-
-		if (targetForm === undefined && targetListData) {
-			// Still no target form. Take the list's current form
-			targetForm = targetListData.form
-			targetGlyph = targetListData.glyph
-		}
-
-		const split = splitCheckboxGlyphs(targetGlyph)
-		targetGlyph = split.base
-		const hasCheckbox = split.box !== undefined
-
-		function getTargetGlyph() {
-			const targetDelimiter: string = getDelimiterForGlyph(targetGlyph)
-			return getGlyphForNumber(targetForm, basisNumber, targetDelimiter) ?? targetGlyph
-		}
-
-		let offset = 0
-		function offsetSelection(position: number, addition: number) {
-			if (selection) {
-				if (selection[0] + offset > position) {
-					selection[0] += addition
-				}
-				if (selection[1] + offset > position) {
-					selection[1] += addition
-				}
-			}
-
-			offset += addition
-		}
-
-		let didSomething = false
-
-		function enforceGlyphOnLine(listData: ListDefinition, lineStart: number, lineText: string) {
-			const { base, box } = splitCheckboxGlyphs(listData.glyph)
-			const targetBase = getTargetGlyph()
-
-			let applyChange = false
-			if (base !== targetBase) {
-				if (options.normalizeUnorderedGlyphs) {
-					applyChange = true
-				}
-				else if (targetForm !== listData.form || !(targetForm === ListForm.Unordered || targetForm === ListForm.UnorderedLarge)) {
-					applyChange = true
-				}
-			}
-			else if (!box && hasCheckbox) {
-				applyChange = true
-			}
-
-			if (applyChange) {
-				change = change || editor.change
-
-				let finalTarget = targetBase
-				if (box) finalTarget += ' ' + box
-				else if (hasCheckbox) finalTarget += ' [ ]'
-
-				const listMatch = lineText.match(listMatcher)
-				const insertedText = listMatch[1] + finalTarget + ' '
-				const sizeDiff = insertedText.length - listMatch[0].length
-				const deleteEnd = lineStart + listMatch[0].length
-				change
-					.delete([lineStart, deleteEnd])
-					.insert(deleteEnd, insertedText)
-
-				didSomething = true
-				
-				offsetSelection(lineStart, sizeDiff)
-				return true
-			}
-			return false
-		}
-
-		// Propagate the target form & basis to the indicated line
-		if (targetListData && targetIndent === intendedIndent) {
-			enforceGlyphOnLine(targetListData, targetRange[0], lineText)
-			if (basisNumber) basisNumber++
-		}
-
-		// Propagate the target form & basis all following list lines on the indent level
-		for (let lineIndex = targetLineIndex + 1; lineIndex < doc.lines.length; lineIndex++) {
-			let nextLine = doc.lines[lineIndex]
-			let nextText = deltaToText(nextLine.content)
-			let indent = nextText.match(indentMatcher)[0]
-
-			if (indent === intendedIndent) {
-				const listData = nextLine.attributes.list as ListDefinition
-				if (!listData) break // End of the road
-				const lineRange = doc.getLineRange(nextLine)
-				enforceGlyphOnLine(listData, lineRange[0], nextText)
-				if (basisNumber) basisNumber++
-			}
-			else if (indent.length < intendedIndent.length) {
-				// We've reached the end of the current indentation.
-				break
-			}
-		}
-
-		if (didSomething && selection && change) {
-			change.select(selection as EditorRange)
-		}
-
-		return change
 	}
 
 	function onChanging(event: EditorChangeEvent) {
